# [WIP] Statement scanning performance improvements