## What does this PR do?

[//]: # (Required: Describe the effects of your pull request in detail. If
multiple changes are involved, a bulleted list is often useful.)

Improves policy & statement evaluation performance by:
- removing redundant calls
- simplifying conditionals and exiting early wherever possible
- adding `cache-property` package so that expensive class attr calls are cached after the first call

Basic testing using `time` and a sample script (pasted below) yielded these results for a policy with a single wildcard statement:
```
    "Statement": [
        {
            "Effect": "Allow",
            "Action": [
                "s3:*"
            ],
            "Resource": "*"
        }
    ]
```

**Before**
```
$ python3 profile_script.py  
1.32s user 0.07s system 99% cpu 1.400s total
```

**After**
```
$ python3 profile_script.py  
0.42s user 0.08s system 92% cpu 0.532 total
```

---
Sample script used to profile:
```
# profile_script.py

def scan_policy(policy_json, exclusions_config=DEFAULT_EXCLUSIONS_CONFIG):
    policy_document = PolicyDocument(policy_json)
    exclusions = Exclusions(DEFAULT_EXCLUSIONS_CONFIG)
    policy_finding = PolicyFinding(policy_document, exclusions)
    return policy_finding.results

if __name__ == '__main__':
    test_filepath = "" # replace with path to a policy file
    with open(test_filepath) as f:
        contents = f.read()
    account_authorization_details_cfg = json.loads(contents)
    rendered_html_report = scan_policy(account_authorization_details_cfg)
```

## What gif best describes this PR or how it makes you feel?

[//]: # (Encouraged: Insert an appropriate gif/meme to brighten up your reviewer's day.)

Still thinking of one :) 

## Completion checklist

- [ ] Additions and changes have unit tests
- [ ] Python Unit tests, Pylint, security testing, and Integration tests are passing.
- [x] Javascript tests are passing (`npm test`)
- [x] If the UI contents or JavaScript files have been modified, generate a new example report (`npm build` and `python3 ./utils/generate_example_report.py`)
- [ ] The pull request has been appropriately labeled using the provided PR labels